widget/tests/test_bug538242.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=538242
-->
<window title="Mozilla Bug 538242"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

<body  xmlns="http://www.w3.org/1999/xhtml">
<div id="content" style="display: none">
  
</div>
<pre id="test">
</pre>
</body>

<script class="testbody" type="application/javascript">
<![CDATA[

if (navigator.platform.startsWith("Win")) {
  SimpleTest.expectAssertions(0, 1);
}

SimpleTest.waitForExplicitFinish();

SimpleTest.waitForFocus(function () {
  if (navigator.platform.includes("Lin")) {
    ok(true, "This test is disabled on Linux because it expects moving windows to be synchronous which is not guaranteed on Linux.");
    SimpleTest.finish();
    return;
  }

  var win = window.open("window_bug538242.xul", "_blank", 
                        "chrome=1,width=400,height=300,left=100,top=100");
  SimpleTest.waitForFocus(function () {
    is(win.screenX, 100, "window should open at 100, 100");
    is(win.screenY, 100, "window should open at 100, 100");
    var [oldX, oldY] = [win.screenX, win.screenY];
    win.moveTo(0, 0);
    isnot(win.screenX, oldX, "window should have moved to a point near 0, 0");
    isnot(win.screenY, oldY, "window should have moved to a point near 0, 0");
    win.close();
    SimpleTest.finish();
  }, win);
});

]]>
</script>

</window>
